feat(io): add CMake build system and CSV data exporter
- Add CMakeLists.txt with C++20 requirement and GLM FetchContent.
- Create DataExporter to serialize graph state and metrics to CSV.
- Update main.cpp to run simulation and save outputs.
- Add Python plotting script for thesis visualization.

+++ Include/graph.hpp
@@ -126,7 +126,7 @@ public:
         for (std::size_t i = 0; i < n; ++i)
             g.addVertex(static_cast<Node::ID>(i));
 
-        std::mt19937_64 rng{ seed.value_or(std::random_device{}()) };
+        std::mt19937_64 rng{ static_cast<std::uint64_t>(seed.value_or(std::random_device{}())) };
         std::bernoulli_distribution coin{ p };
 
         for (std::size_t i = 0; i < n; ++i)
